+++ crates/memofs/src/lib.rs
@@ -190,7 +190,7 @@ impl VfsInner {
     fn read_raw(&mut self, path: &Path) -> io::Result<Vec<u8>> {
         if let Some(cache) = &mut self.prefetch_cache {
             if let Some(contents) = cache.files.remove(path) {
-                if self.watch_enabled && self.watch_recursive {
+                if self.watch_enabled {
                     self.watch_or_record(path)?;
                 }
                 return Ok(contents);
@@ -199,7 +199,7 @@ impl VfsInner {
 
         let contents = self.backend.read(path)?;
 
-        if self.watch_enabled && self.watch_recursive {
+        if self.watch_enabled {
             self.watch_or_record(path)?;
         }